+++ CHANGELOG.md
@@ -24,6 +24,10 @@ Line wrap the file at 100 chars.                                              Th
 ## [Unreleased]
 
 
+## [2026.5] - 2026-09-11
+This release is identical to 2026.5-beta2.
+
+
 ## [2026.5-beta2] - 2026-09-08
 ### Changed
 - Do not show dialog when rendering the map fails due to disabled GPU acceleration.